+++ packages/tasq-server/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./migration.js";
 export * from "./store.js";
 export * from "./router.js";
+export * from "./http.js";